refact(nftType): it update the type of the variable nftType from uint8 to uint16, and idToken from uint8 to uint32;

+++ contracts/structs/TokenData.sol
@@ -2,6 +2,6 @@
 pragma solidity ^0.8.20;
 
 struct TokenData {
-    uint8 nftType;
+    uint16 nftType;
     address royaltyAddress;
 }
